+++ InfiniteRecharge/src/main/java/frc/robot/subsystem/drive/DriveConstants.java
@@ -10,21 +10,10 @@ package frc.robot.subsystem.drive;
 import com.ctre.phoenix.motorcontrol.FeedbackDevice;
 
 import edu.wpi.first.wpilibj.kinematics.DifferentialDriveKinematics;
-import frc.robot.config.Config;
 /**
  * Add your docs here.
  */
 public class DriveConstants {
-    private static Config config;
-    public static void setConfig(Config c) {
-        config = c;
-
-        WHEEL_CIRCUMFERENCE_INCHES = 2*Math.PI*c.drive.wheelRadius_in;
-        KINEMATICS = new DifferentialDriveKinematics(c.drive.trackWidth_in * METERS_PER_INCH);
-    }
-    
-
-
     public static final double JOYSTICK_DEADBAND = 0.15;
 
     // Set velocity follower type to false when independent gear boxes are being used
@@ -45,32 +34,11 @@ public class DriveConstants {
     public static final double STANDARD_G_FTPSPS = 32.1740;
     public static final double MAX_LAT_ACCELERATION_IPSPS = STANDARD_G_FTPSPS * 12.0;
 
-    // TODO
-    // defined when given a config
-    public static double WHEEL_CIRCUMFERENCE_INCHES;
-
     // Identify what type of feedback device we will use on this drive base
     // Assume that all feedback devices are the same type on all axels that
     // need to be measured.
     // TODO
     public static final FeedbackDevice DRIVE_MOTOR_FEEDBACK_DEVICE = FeedbackDevice.QuadEncoder;
-
-    // The motor controllers we use (TalonSRX) return velocity in terms of native ticks per 100 ms 
-    // and expect commands to be similarly dimensioned.
-    // Even though this is a constants package, we provide the convenient conversion to/from
-    // inches per second to the native ticks per 100 ms.
-    // NOTE: Integer truncation is assumed for a maximum reduction of 10 ticks per second.
-    // For 8192 ticks per rev that is error of < 0.07 RPM
-    public static int ipsToTicksP100(double ips)
-    {
-        double rps = ips / WHEEL_CIRCUMFERENCE_INCHES;
-        return (int) (config.drive.ticksPerRevolution * rps / 10.0);
-    }
-    public static double ticksP100ToIps(int ticksP100)
-    {
-        double rps = ticksP100 * 10.0 / (config.drive.ticksPerRevolution);
-        return rps * WHEEL_CIRCUMFERENCE_INCHES;
-    }
 
     // FIRST !!! get the sensor phase correct.
     // If positive input to motor controller (green LED) makes the sensor
